VotingPlatform/requirements.txt
amqp==2.2.2
anyjson==0.3.3
appdirs==1.4.3
billiard==3.3.0.23
mysqlclient==1.3.12
Django==2.0.2
django-bootstrap3==9.1.0
django-celery==3.2.2
django-sendfile==0.3.11
django-extensions==2.0.6
pygraphviz==1.3.1
kombu==4.1.0
packaging==16.8
pyparsing==2.2.0
pytz==2018.3
six==1.11.0
vine==1.1.4
